+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Component.log() now includes the date and [SDK] prefix.
 
 ## [1.2.1] - 2017-10-01
 ## Changed