+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Runtime call default timeout to 10000.
+- Parameter schema default value getter now returns None by default.
 
 ### Fixed
 - Component.log() now includes the date and [SDK] prefix.